app/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
requests==2.28.1
urllib3==1.26.11
vk-api==11.9.9
pydantic==1.9.2
psycopg2==2.9.3
sqlalchemy==1.4.40
greenlet==1.1.2
py3-validate-email==1.0.7
schedule==1.1.0
